refactor(contracts): move IPC response decoders beside the contract
