+++ packages/contracts/src/ipc.ts
@@ -2,6 +2,7 @@ export * from "./ipc-app-auth";
 export * from "./ipc-browser";
 export * from "./ipc-channels";
 export * from "./ipc-conversation";
+export * from "./ipc-decoders";
 export * from "./ipc-desktop-apis";
 export * from "./ipc-dynamic-island";
 export * from "./ipc-hosted-sites";
